@@ -1,6 +1,6 @@
 {
   "name": "eventsource",
-  "version": "1.1.0",
+  "version": "2.0.2",
   "description": "W3C compliant EventSource client for Node.js and browser (polyfill)",
   "keywords": [
     "eventsource",
@@ -46,14 +46,15 @@
     "coverage": "nyc --reporter=html --reporter=text _mocha --reporter spec"
   },
   "engines": {
-    "node": ">=0.12.0"
+    "node": ">=12.0.0"
   },
-  "dependencies": {
-    "original": "^1.0.0"
-  },
+  "dependencies": {},
   "standard": {
     "ignore": [
       "example/eventsource-polyfill.js"
+    ],
+    "globals": [
+      "URL"
     ]
   }
 }
