@@ -1,9 +1,13 @@
 {
   "name": "sockjs-client",
   "description": "SockJS-client is a browser JavaScript library that provides a WebSocket-like object.",
-  "version": "1.5.2",
+  "version": "1.6.1",
   "author": "Bryce Kahle",
   "jsdelivr": "dist/sockjs.min.js",
+  "engines": {
+    "node": ">=12"
+  },
+  "funding": "https://tidelift.com/funding/github/npm/sockjs-client",
   "browser": {
     "./lib/transport/driver/websocket.js": "./lib/transport/browser/websocket.js",
     "eventsource": "./lib/transport/browser/eventsource.js",
@@ -25,34 +29,33 @@
     }
   ],
   "dependencies": {
-    "debug": "^3.2.6",
-    "eventsource": "^1.0.7",
-    "faye-websocket": "^0.11.3",
+    "debug": "^3.2.7",
+    "eventsource": "^2.0.2",
+    "faye-websocket": "^0.11.4",
     "inherits": "^2.0.4",
-    "json3": "^3.3.3",
-    "url-parse": "^1.5.3"
+    "url-parse": "^1.5.10"
   },
   "devDependencies": {
-    "browserify": "^16.5.1",
+    "browserify": "^17.0.0",
     "envify": "^4.0.0",
-    "eslint": "^7.6.0",
+    "eslint": "^8.10.0",
     "expect.js": "~0.3.1",
     "gulp": "^4.0.2",
     "gulp-header": "^2.0.9",
     "gulp-rename": "^2.0.0",
-    "gulp-replace": "^1.0.0",
-    "gulp-sourcemaps": "^2.6.5",
+    "gulp-replace": "^1.1.3",
+    "gulp-sourcemaps": "^3.0.0",
     "gulp-uglify": "^3.0.2",
-    "karma": "^5.1.1",
-    "karma-browserify": "^7.0.0",
+    "karma": "^6.3.16",
+    "karma-browserify": "^8.1.0",
     "karma-browserstack-launcher": "^1.6.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^2.0.1",
-    "mocha": "^8.1.0",
+    "mocha": "^9.2.1",
     "proxyquire": "^2.1.3",
     "pump": "^3.0.0",
-    "serve-static": "^1.14.1",
-    "sockjs": "^0.3.21",
+    "serve-static": "^1.14.2",
+    "sockjs": "^0.3.24",
     "vinyl-buffer": "~1.0.0",
     "vinyl-source-stream": "^2.0.0"
   },
