@@ -89,39 +89,40 @@ export default function JadwalPelajaranAdmin() {
                 onBack={() => window.history.back()}
                 title="Jadwal Pelajaran"
             />
-            <Card className="w-full shadow-md py-0 mb-4">
-              <Row>
-                <Col span={12}>
-                  <Button onClick={() => setIsViewPelajaran(false)} className="mr-2 bg-current" type="primary" shape="round" size='middle'>
+            <Card className="card bg-lightblue border-0 mb-4 text-grey-900">
+              <div className="row">
+                <div className="col-lg-8 col-md-6 my-2">
+                  <Button className="mr-4" type="primary" shape="round" size='middle'
+                          onClick={() => setIsViewPelajaran(false)}>
                     Tambah Data
                   </Button>
                   <Dropdown overlay={_filterMenu}>
-                    <a className="ant-dropdown-link mr-2 font-bold" onClick={e => e.preventDefault()}>
+                    <a className="ant-dropdown-link mr-4 font-bold"
+                    onClick={e => e.preventDefault()}>
                       Filter by <DownOutlined/>
                     </a>
                   </Dropdown>
                   <Dropdown overlay={_sortMenu}>
-                    <a className="ant-dropdown-link font-bold" onClick={e => e.preventDefault()}>
+                    <a className="ant-dropdown-link font-bold"
+                    onClick={e => e.preventDefault()}>
                       Sort by <DownOutlined/>
                     </a>
                   </Dropdown>
-                </Col>
-                <Col span={12}>
-                  <div className="float-right">
-                    <Search className="mr-6" placeholder="Cari kata kunci" allowClear
-                            onSearch={_onSearch} style={{width: 250, lineHeight: '20px'}}/>
-                    {grid == false ? 
+                </div>
+                <div className="col-lg-4 col-md-6 my-2">
+                  <Search className="mr-3" placeholder="Cari kata kunci" allowClear
+                          onSearch={_onSearch} style={{width: '80%'}}/>
+                  {grid == false ?
                     <a>
-                    <AppstoreOutlined style={{fontSize: '30px'}}
-                                        onClick={() => setGrid(true)}/>
+                      <AppstoreOutlined style={{fontSize: '2em', lineHeight: 1}}
+                                      onClick={() => setGrid(true)}/>
                     </a> :
                     <a>
-                    <MenuOutlined style={{fontSize: '30px'}}
-                                    onClick={() => setGrid(false)}/>
+                      <MenuOutlined style={{fontSize: '2em', lineHeight: 1}}
+                                  onClick={() => setGrid(false)}/>
                     </a>}
-                  </div>
-                </Col>
-              </Row>
+                </div>
+              </div>
             </Card>
             <div className="px-1 py-2 ">
               <div className="row">
